추가한 코드 MainActivity에서 specificroute로 이동
